Update docker-container-actions.ts
Hi,
When you run or exec a container, you can pass command argument to the command line. the official api takes an array of string in arguments (see cmd argument reference *)
In your implementation you are assuming that every command should be run via sh -c and so you are building a command line with "sh -c"
even if it is easier for the consumer, as a string can be used with space in it, it does not comply with docker api and it is not as flexible as desired : if the container does not embed sh, it will fail to execute.
Proposition is made to change the variable type from string to array of string. this will allow to use the full power of docker run/exec.

Exemple with rancher/cli2 container : 

cmd = "--version"

$ docker run rancher/cli2 sh -c '--version'
No help topic for 'sh'

$ docker run rancher/cli2 --version
rancher version v2.7.0




Be aware that it is a breaking change for existing flows.



*https://docs.docker.com/engine/api/v1.41/#tag/Exec/operation/ContainerExec
*https://docs.docker.com/engine/api/v1.41/#tag/Container/operation/ContainerCreate